requirements.txt
# Core dependencies
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.21.0

# Document processing
pypdf2>=3.0.0
python-docx==0.8.11

# Google Cloud Services
google-cloud-vision>=3.0.0
google-generativeai>=0.3.0

# Local model support
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
accelerate>=0.20.0
